convert logical to int8
svn path=/trunk/matlab/; revision=8074


Former-commit-id: 394f237ba116914f19195e89fdeffe3983cf0114

+++ applications/Rijkswaterstaat/jarkus/jarkus_transect2netcdf.m
@@ -92,7 +92,7 @@ for i = 1 : length(yearArray)
     % should this be in jarkus_transect
     nc_varput(filename, 'min_cross_shore_measurement', minCrossBlock , [i-1, 0], [1, length(minCrossBlock)])
     nc_varput(filename, 'max_cross_shore_measurement', maxCrossBlock , [i-1, 0], [1, length(maxCrossBlock)])
-    nc_varput(filename, 'has_data',            ~isnan(maxCrossBlock) , [i-1, 0], [1, length(maxCrossBlock)])
+    nc_varput(filename, 'has_data',      int8(~isnan(maxCrossBlock)) , [i-1, 0], [1, length(maxCrossBlock)])
     nc_varput(filename, 'min_altitude_measurement', minAltitBlock , [i-1, 0], [1, length(minAltitBlock)])
     nc_varput(filename, 'max_altitude_measurement', maxAltitBlock , [i-1, 0], [1, length(minAltitBlock)])
     nc_varput(filename, 'time'      , year          , [i-1      ], [1                        ]);
